SOL-151648: Split executor tests into separate file
